Cleaned up Thrust handler slightly
Switched to a delayed hit design like the monster skill handler, which seems to have fixed the early damage animation.

+++ src/ZoneServer/World/Actors/Characters/Character.cs
@@ -699,9 +699,9 @@ namespace Melia.Zone.World.Actors.Characters
 		/// client with ZC_UPDATE_SP.
 		/// </summary>
 		/// <param name="amount"></param>
-		public void ModifySp(int amount)
+		public void ModifySp(float amount)
 		{
-			var sp = (int)this.Properties.Modify(PropertyName.SP, amount);
+			var sp = this.Properties.Modify(PropertyName.SP, amount);
 			Send.ZC_UPDATE_SP(this, sp, true);
 		}
